Thu, Oct 19, 2023, 9:37 PM +03:00

+++ readability/readability.c
@@ -32,7 +32,7 @@ int main(void)
     float L = (float) (100.0 * letters) /((float) words);
     float S = (float) (100.0 * sentances) /((float) words);
 
-    int index = (0.0588 * L) - (0.296 * S) - 15.8;
+    float index = (0.0588 * L) - (0.296 * S) - 15.8;
 
     if (index < 1)
     {
